+++ lib/src/db/postgresql/predicate_builder.dart
@@ -21,56 +21,52 @@ abstract class PredicateBuilder implements EntityTableMapper {
     return obj.backingMap.keys
         .map((propertyName) {
           var desc = obj.entity.properties[propertyName];
-          if (desc is ManagedRelationshipDescription) {
-            if (desc.relationshipType == ManagedRelationshipType.belongsTo) {
+          var innerMatcher = obj.backingMap[propertyName];
+
+          if (desc is ManagedAttributeDescription) {
+            return [
+              new PropertyExpression(this, obj.entity.properties[propertyName],
+                  obj.backingMap[propertyName],
+                  additionalVariablePrefix: prefix)
+            ];
+          }
+
+          // But if it is a relationship and only the primary key value is set, then we can do t
+          // this without the join.. its just the foreign key column
+          var relationshipDesc = desc as ManagedRelationshipDescription;
+          if (relationshipDesc.relationshipType == ManagedRelationshipType.belongsTo) {
+            var innerMatcherObject = innerMatcher as ManagedObject;
+            var nestedMatcherPrimaryKeyMatcher = innerMatcherObject.backingMap[innerMatcherObject.entity.primaryKey];
+            if (innerMatcherObject.backingMap.length == 1 && nestedMatcherPrimaryKeyMatcher != null) {
               return [
-                new PropertyExpression(
-                    this,
-                    obj.entity.properties[propertyName],
-                    obj.backingMap[propertyName],
+                new PropertyExpression(this, desc, nestedMatcherPrimaryKeyMatcher,
                     additionalVariablePrefix: prefix)
               ];
             }
-
-            // Otherwise, this is an implicit join...
-            // Do we have an existing guy?
-            bool disambiguate = true;
-            RowMapper innerRowMapper = returningOrderedMappers
-                .where((m) => m is RowMapper)
-                .firstWhere(
-                    (m) => (m as RowMapper).representsRelationship(desc),
-                    orElse: () => null);
-
-            if (innerRowMapper == null) {
-              innerRowMapper =
-                  new RowMapper.implicit(PersistentJoinType.leftOuter, desc);
-              innerRowMapper.parentTable = this;
-              createdImplicitRowMappers.add(innerRowMapper);
-              disambiguate = false;
-            }
-
-            var innerMatcher = obj.backingMap[propertyName];
-            if (innerMatcher is NullMatcherExpression) {
-              innerMatcher = new ManagedObject()
-                ..entity = desc.inverseRelationship.entity
-                ..backing = new ManagedMatcherBacking()
-                ..[desc.inverseRelationship.name] = innerMatcher;
-            }
-
-            if (innerMatcher is ManagedSet) {
-              innerMatcher = (innerMatcher as ManagedSet).matchOn;
-            }
-
-            return innerRowMapper.propertyExpressionsFromObject(
-                innerMatcher, createdImplicitRowMappers,
-                disambiguateVariableNames: disambiguate);
           }
 
-          return [
-            new PropertyExpression(this, obj.entity.properties[propertyName],
-                obj.backingMap[propertyName],
-                additionalVariablePrefix: prefix)
-          ];
+          bool disambiguate = true;
+          RowMapper innerRowMapper = returningOrderedMappers
+              .where((m) => m is RowMapper)
+              .firstWhere(
+                  (m) => (m as RowMapper).representsRelationship(desc),
+                  orElse: () => null);
+
+          if (innerRowMapper == null) {
+            innerRowMapper =
+                new RowMapper.implicit(PersistentJoinType.leftOuter, desc);
+            innerRowMapper.originatingTable = this;
+            createdImplicitRowMappers.add(innerRowMapper);
+            disambiguate = false;
+          }
+
+          if (innerMatcher is ManagedSet) {
+            innerMatcher = (innerMatcher as ManagedSet).matchOn;
+          }
+
+          return innerRowMapper.propertyExpressionsFromObject(
+              innerMatcher, createdImplicitRowMappers,
+              disambiguateVariableNames: disambiguate);
         })
         .expand((expressions) => expressions)
         .toList();